+++ source/extensions/isaacsim.util.internal/isaacsim/util/internal/extension.py
@@ -104,10 +104,7 @@ class InternalTools(omni.ext.IExt):
 
     def get_unreleasable(self):
         asset_paths = [
-            "/Isaac/Robots/UR10/robotiq",
-            "/Isaac/Robots/UR10/ur10_robotiq.usd",
-            "/Isaac/Robots/UR10/ur10_schmalz.usd",
-            "/Isaac/Samples/Leonardo/Stage/ur10_bin_stacking_robotiq.usd",
+            # Isaac Sim does not have any unreleasable assets in 5.0.0
         ]
 
         async def check():